Cleaned Up For Presentation
Moved BERT stuff to dedicated file mostly. Light cleanup of main
